Moved inline JS to its own file and enqueued it
