+++ duplicate-post-options.php
@@ -6,6 +6,8 @@
  * @since   2.0
  */
 
+use Yoast\WP\Duplicate_Post\Utils;
+
 if ( ! defined( 'ABSPATH' ) ) {
 	exit();
 }
@@ -126,16 +128,35 @@ function duplicate_post_menu() {
 		'duplicatepost',
 		'duplicate_post_options'
 	);
-	add_action( $page_hook, 'duplicate_post_add_options_page_css' );
+	add_action( $page_hook, 'duplicate_post_add_options_page_assets' );
 }
 
 /**
- * Enqueues a CSS file with styles for the options page.
+ * Enqueues assets for the options page.
  *
  * @ignore
  */
-function duplicate_post_add_options_page_css() {
-	wp_enqueue_style( 'duplicate-post-options', plugins_url( '/duplicate-post-options.css', __FILE__ ), array(), DUPLICATE_POST_CURRENT_VERSION );
+function duplicate_post_add_options_page_assets() {
+	\wp_enqueue_style(
+	        'duplicate-post-options',
+            \plugins_url( '/duplicate-post-options.css', __FILE__ ),
+            [],
+            DUPLICATE_POST_CURRENT_VERSION
+    );
+
+	\wp_enqueue_script(
+	        'duplicate_post_options_script',
+            \plugins_url(
+                    \sprintf(
+                            'js/dist/duplicate-post-options-%s.js',
+                            Utils::flatten_version( DUPLICATE_POST_CURRENT_VERSION )
+                    ),
+                    DUPLICATE_POST_FILE
+            ),
+            [ 'jquery' ],
+            DUPLICATE_POST_CURRENT_VERSION,
+            true
+	);
 }
 
 function generate_options_checkbox_list( $options ) {
@@ -251,189 +272,6 @@ function duplicate_post_options() {
         <h1>
 			<?php esc_html_e( 'Duplicate Post Options', 'duplicate-post' ); ?>
         </h1>
-
-        <script>
-			var tablist;
-			var tabs;
-			var panels;
-
-			// For easy reference
-			var keys = {
-				end: 35,
-				home: 36,
-				left: 37,
-				up: 38,
-				right: 39,
-				down: 40,
-				delete: 46
-			};
-
-			// Add or substract depending on key pressed
-			var direction = {
-				37: - 1,
-				38: - 1,
-				39: 1,
-				40: 1
-			};
-
-
-			function generateArrays() {
-				tabs = document.querySelectorAll( "#duplicate_post_settings_form [role=\"tab\"]" );
-				panels = document.querySelectorAll( "#duplicate_post_settings_form [role=\"tabpanel\"]" );
-			}
-
-			function addListeners( index ) {
-				tabs[index].addEventListener( "click", function ( event ) {
-					var tab = event.target;
-					activateTab( tab, false );
-				} );
-				tabs[index].addEventListener( "keydown", function ( event ) {
-					var key = event.keyCode;
-
-					switch ( key ) {
-						case keys.end:
-							event.preventDefault();
-							// Activate last tab
-							activateTab( tabs[tabs.length - 1] );
-							break;
-						case keys.home:
-							event.preventDefault();
-							// Activate first tab
-							activateTab( tabs[0] );
-							break;
-					}
-				} );
-				tabs[index].addEventListener( "keyup", function ( event ) {
-					var key = event.keyCode;
-
-					switch ( key ) {
-						case keys.left:
-						case keys.right:
-							switchTabOnArrowPress( event );
-							break;
-					}
-				} );
-
-				// Build an array with all tabs (<button>s) in it
-				tabs[index].index = index;
-			}
-
-
-			// Either focus the next, previous, first, or last tab
-			// depening on key pressed
-			function switchTabOnArrowPress( event ) {
-				var pressed = event.keyCode;
-
-				for ( x = 0; x < tabs.length; x ++ ) {
-					tabs[x].addEventListener( "focus", focusEventHandler );
-				}
-
-				if ( direction[pressed] ) {
-					var target = event.target;
-					if ( target.index !== undefined ) {
-						if ( tabs[target.index + direction[pressed]] ) {
-							tabs[target.index + direction[pressed]].focus();
-						}
-						else if ( pressed === keys.left || pressed === keys.up ) {
-							focusLastTab();
-						}
-						else if ( pressed === keys.right || pressed === keys.down ) {
-							focusFirstTab();
-						}
-					}
-				}
-			}
-
-			// Activates any given tab panel
-			function activateTab( tab, setFocus ) {
-				setFocus = setFocus || true;
-				// Deactivate all other tabs
-				deactivateTabs();
-
-				// Remove tabindex attribute
-				tab.removeAttribute( "tabindex" );
-
-				// Set the tab as selected
-				tab.setAttribute( "aria-selected", "true" );
-
-				tab.classList.add( "nav-tab-active" );
-
-				// Get the value of aria-controls (which is an ID)
-				var controls = tab.getAttribute( "aria-controls" );
-
-				// Remove hidden attribute from tab panel to make it visible
-				document.getElementById( controls )
-				        .removeAttribute( "hidden" );
-
-				// Set focus when required
-				if ( setFocus ) {
-					tab.focus();
-				}
-			}
-
-			// Deactivate all tabs and tab panels
-			function deactivateTabs() {
-				for ( t = 0; t < tabs.length; t ++ ) {
-					tabs[t].setAttribute( "tabindex", "-1" );
-					tabs[t].setAttribute( "aria-selected", "false" );
-					tabs[t].classList.remove( "nav-tab-active" );
-					tabs[t].removeEventListener( "focus", focusEventHandler );
-				}
-
-				for ( p = 0; p < panels.length; p ++ ) {
-					panels[p].setAttribute( "hidden", "hidden" );
-				}
-			}
-
-			// Make a guess
-			function focusFirstTab() {
-				tabs[0].focus();
-			}
-
-			// Make a guess
-			function focusLastTab() {
-				tabs[tabs.length - 1].focus();
-			}
-
-			//
-			function focusEventHandler( event ) {
-				var target = event.target;
-
-				checkTabFocus( target );
-			}
-
-			// Only activate tab on focus if it still has focus after the delay
-			function checkTabFocus( target ) {
-				focused = document.activeElement;
-
-				if ( target === focused ) {
-					activateTab( target, false );
-				}
-			}
-
-			document.addEventListener( "DOMContentLoaded", function () {
-				tablist = document.querySelectorAll( "#duplicate_post_settings_form [role=\"tablist\"]" )[0];
-
-				generateArrays();
-
-				// Bind listeners
-				for ( i = 0; i < tabs.length; ++ i ) {
-					addListeners( i );
-				}
-			} );
-
-		jQuery(function(){
-			jQuery('.taxonomy_private').hide();
-
-			jQuery( '.toggle-private-taxonomies' )
-				.on( 'click', function() {
-					buttonElement = jQuery( this );
-					jQuery( '.taxonomy_private' ).toggle( 300, function() {
-						buttonElement.attr( 'aria-expanded', jQuery( this ).is( ":visible" ) );
-					} );
-				} );
-		});
-	</script>
 
         <form id="duplicate_post_settings_form" method="post" action="options.php" style="clear: both">
 			<?php settings_fields( 'duplicate_post_group' ); ?>
@@ -851,7 +689,7 @@ function duplicate_post_options() {
                         <th scope="row"><?php esc_html_e( 'Show these links', 'duplicate-post' ); ?></th>
                         <td>
                             <fieldset>
-								<?php generate_options_checkbox_list( $link_display_options ); ?>
+								<?php echo generate_options_checkbox_list( $link_display_options ); ?>
                             </fieldset>
                         </td>
                     </tr>
@@ -860,7 +698,7 @@ function duplicate_post_options() {
                         <th scope="row"><?php esc_html_e( 'Show links in', 'duplicate-post' ); ?></th>
                         <td>
                             <fieldset>
-								<?php generate_options_checkbox_list( $link_location_options ); ?>
+								<?php echo generate_options_checkbox_list( $link_location_options ); ?>
                             </fieldset>
                             <p>
 								<?php esc_html_e( 'Whether the links are displayed for custom post types registered by themes or plugins depends on their use of standard WordPress UI elements.', 'duplicate-post' ); ?>
